exprstress: A randomized stress tester for expression evaluation
Although we have a fair number of hand-written unit tests for different
permutations of expression evaluation, when writing those we tend to
favor relatively straightforward situations which only include a single
operator or function at a time, and thus we can sometimes miss including
test cases for less common situations that can arise when combining
different features together.

This new package and its associated CLI tool are intended to randomly
generate various valid expressions with different combinations of
expression operators and functions from the Terraform language, and then
check to make sure that they evaluate successfully and produce a result
of the expected type, mode, and sensitivity. Hopefully this will allow us
to find problems that we wouldn't think to test.

This is a similar sort of idea as fuzz testing, but at a higher level of
abstraction where the input is intended to always be a valid expression
and we're focused on the evaluator, rather than generating random bytes
and trying to crash the HCL parser. This is a form of property-based
testing, albeit applying the same set of properties to all possible
expressions rather than testing unique properties of each expression type.